build 1.2.2 patch with custom JACK lookup
In version 1.2.2 I introduced a new compile option `WANT_DYNAMIC_JACK_CHECK` which makes Hydrogen check during startup whether either `jackd`, `jackdbus`, or `pw-jack` is present and enables/disables JACK support accordingly.

But I totally forgot that the Flatpak just works with Pipewire and not the classical `jackd`. Damn. I made a small patch and introduced an additional option `WANT_PIPEWIRE_JACK_ONLY_CHECK`. It is not on a tag anymore but I hope this is still okay.